Rename XetContext -> XetRuntime, XetRuntime -> XetThreadpool
Pull config and common out of the threadpool into the runtime context.
Rename the runtime field to threadpool. Add factory methods on XetRuntime
that delegate to XetThreadpool for construction.

Made-with: Cursor
