+++ git_xet/src/token_refresher.rs
@@ -4,7 +4,7 @@ use http::header::HeaderMap;
 use xet_client::cas_client::auth::DirectRefreshRouteTokenRefresher;
 use xet_client::common::http_client::build_http_client;
 use xet_client::hub_client::Operation;
-use xet_runtime::core::XetContext;
+use xet_runtime::core::XetRuntime;
 
 use crate::auth::get_credential;
 use crate::errors::Result;
@@ -14,7 +14,7 @@ use crate::git_url::GitUrl;
 /// Build a [`DirectRefreshRouteTokenRefresher`] for the git-xet path,
 /// deriving credentials from the git repo's credential helper.
 pub fn new_git_token_refresher(
-    ctx: &XetContext,
+    ctx: &XetRuntime,
     repo: &GitRepo,
     remote_url: Option<GitUrl>,
     refresh_route: &str,
