+++ hf_xet/src/lib.rs
@@ -20,7 +20,7 @@ use tracing::debug;
 use xet_pkg::XetError;
 use xet_pkg::legacy::progress_tracking::TrackingProgressUpdater;
 use xet_pkg::legacy::{Sha256Policy, XetFileInfo, data_client};
-use xet_runtime::core::{XetContext, file_handle_limits};
+use xet_runtime::core::{XetRuntime, file_handle_limits};
 
 use crate::logging::init_logging;
 use crate::progress_update::WrappedProgressUpdater;
@@ -328,7 +328,7 @@ pub fn force_sigint_shutdown() -> PyResult<()> {
 
 fn try_parse_progress_updaters(
     funcs: Vec<Py<PyAny>>,
-    ctx: Arc<XetContext>,
+    ctx: Arc<XetRuntime>,
 ) -> PyResult<Vec<Arc<dyn TrackingProgressUpdater>>> {
     let mut updaters = Vec::with_capacity(funcs.len());
     for updater_func in funcs {
