+++ hf_xet/src/progress_update.rs
@@ -8,7 +8,7 @@ use pyo3::types::{IntoPyDict, PyList, PyString};
 use pyo3::{IntoPyObjectExt, Py, PyAny, PyResult, Python, pyclass};
 use tracing::error;
 use xet_pkg::legacy::progress_tracking::{ProgressUpdate, TrackingProgressUpdater};
-use xet_runtime::core::XetContext;
+use xet_runtime::core::XetRuntime;
 use xet_runtime::error_printer::ErrorPrinter;
 
 use crate::runtime::convert_multithreading_error;
@@ -121,7 +121,7 @@ pub struct PyTotalProgressUpdate {
 /// passed around as a ProgressUpdater trait object or
 /// as a template parameter
 struct WrappedProgressUpdaterImpl {
-    ctx: Arc<XetContext>,
+    ctx: Arc<XetRuntime>,
 
     /// Is this enabled?
     progress_updating_enabled: bool,
@@ -146,7 +146,7 @@ impl Debug for WrappedProgressUpdaterImpl {
 const DETAILED_PROGRESS_ARG_NAMES: [&str; 2] = ["total_update", "item_updates"];
 
 impl WrappedProgressUpdaterImpl {
-    pub fn new(py_func: Py<PyAny>, ctx: Arc<XetContext>) -> PyResult<Self> {
+    pub fn new(py_func: Py<PyAny>, ctx: Arc<XetRuntime>) -> PyResult<Self> {
         // Analyze the function to make sure it's the correct form. If it's 4 arguments with
         // the appropriate names, than we call it using the detailed progress update; if it's
         // a single function, we assume it's a global increment function and just pass in the update
@@ -225,7 +225,7 @@ impl WrappedProgressUpdaterImpl {
     }
 
     async fn register_updates_impl(self: Arc<Self>, updates: ProgressUpdate) -> PyResult<()> {
-        let rt = self.ctx.runtime.clone();
+        let rt = self.ctx.threadpool.clone();
         rt.spawn_blocking(move || {
             Python::attach(|py| {
                 let f = self.py_func.bind(py);
@@ -298,7 +298,7 @@ pub struct WrappedProgressUpdater {
 }
 
 impl WrappedProgressUpdater {
-    pub fn new(py_func: Py<PyAny>, ctx: Arc<XetContext>) -> PyResult<Self> {
+    pub fn new(py_func: Py<PyAny>, ctx: Arc<XetRuntime>) -> PyResult<Self> {
         Ok(Self {
             inner: Arc::new(WrappedProgressUpdaterImpl::new(py_func, ctx)?),
         })
