+++ simulation/src/upload_concurrency/upload_simulation_client.rs
@@ -21,8 +21,7 @@ use xet_client::cas_client::adaptive_concurrency::{
 use xet_client::cas_client::progress_tracked_streams::UploadProgressStream;
 use xet_client::cas_client::retry_wrapper::RetryWrapper;
 use xet_client::common::http_client::build_http_client;
-use xet_runtime::config::XetConfig;
-use xet_runtime::core::{XetContext, XetRuntime};
+use xet_runtime::core::XetRuntime;
 
 use crate::scenario::base_url;
 
@@ -191,7 +190,7 @@ fn spawn_stats_reporter(
 
 /// Shared context for upload workers.
 struct UploadContext {
-    xet_ctx: XetContext,
+    xet_ctx: XetRuntime,
     url: String,
     http_client: reqwest_middleware::ClientWithMiddleware,
     base_data: Bytes,
@@ -315,7 +314,7 @@ async fn run_upload_clients_impl(
     let max_data_size = max_data_kb * 1024;
     let client_id = rand::rng().random_range(0..1000000000_u64);
 
-    let xet_ctx = XetContext::new(XetRuntime::from_external(tokio::runtime::Handle::current()), XetConfig::new());
+    let xet_ctx = XetRuntime::from_external(tokio::runtime::Handle::current());
     let http_client = build_http_client(&xet_ctx, "test_session", None, None)?;
 
     let duration_sec = repeat_duration_seconds.unwrap_or(u64::MAX);
