+++ xet_client/src/cas_client/retry_wrapper.rs
@@ -9,7 +9,7 @@ use tokio::sync::Mutex;
 use tokio_retry::RetryIf;
 use tokio_retry::strategy::{ExponentialBackoff, jitter};
 use tracing::{error, info};
-use xet_runtime::core::XetContext;
+use xet_runtime::core::XetRuntime;
 
 use super::adaptive_concurrency::ConnectionPermit;
 use crate::common::http_client::request_id_from_response;
@@ -28,7 +28,7 @@ struct ConnectionPermitInfo {
 
 pub struct RetryWrapper {
     #[allow(dead_code)]
-    ctx: XetContext,
+    ctx: XetRuntime,
     max_attempts: usize,
     base_delay: Duration,
     no_retry_on_429: bool,
@@ -39,7 +39,7 @@ pub struct RetryWrapper {
 }
 
 impl RetryWrapper {
-    pub fn new(ctx: XetContext, api_tag: &'static str) -> Self {
+    pub fn new(ctx: XetRuntime, api_tag: &'static str) -> Self {
         let max_attempts = ctx.config.client.retry_max_attempts;
         let base_delay = ctx.config.client.retry_base_delay;
         Self {
@@ -541,16 +541,13 @@ mod tests {
     use wiremock::matchers::{method, path};
     use wiremock::{Mock, MockServer, ResponseTemplate};
     use xet_runtime::config::XetConfig;
-    use xet_runtime::core::{XetContext, XetRuntime};
+    use xet_runtime::core::XetRuntime;
 
     use super::*;
 
-    fn test_ctx() -> XetContext {
+    fn test_ctx() -> XetRuntime {
         let config = XetConfig::new();
-        XetContext::new(
-            XetRuntime::from_external_with_config(tokio::runtime::Handle::current(), &config).expect("test ctx"),
-            config,
-        )
+        XetRuntime::from_external_with_config(tokio::runtime::Handle::current(), config).expect("test ctx")
     }
 
     fn connection_wrapper(api: &'static str) -> RetryWrapper {
