+++ xet_client/src/cas_client/simulation/local_server/server.rs
@@ -47,7 +47,7 @@ use tokio::net::TcpListener;
 #[cfg(test)]
 use tokio::sync::oneshot;
 use tower_http::cors::CorsLayer;
-use xet_runtime::core::XetContext;
+use xet_runtime::core::XetRuntime;
 
 #[cfg(test)]
 use super::super::super::RemoteClient;
@@ -107,7 +107,7 @@ impl LocalServer {
     /// If `in_memory` is false, creates a new `LocalClient` pointing to the configured data directory.
     /// If `in_memory` is true, creates a new `MemoryClient` (data directory is ignored).
     pub async fn new(config: LocalServerConfig) -> Result<Self> {
-        let ctx = XetContext::default().map_err(|e| ClientError::Other(e.to_string()))?;
+        let ctx = XetRuntime::default().map_err(|e| ClientError::Other(e.to_string()))?;
         let (client, deletion_client): (Arc<dyn DirectAccessClient>, Option<Arc<dyn DeletionControlableClient>>) =
             if config.in_memory {
                 (MemoryClient::new(ctx.clone()), None)
@@ -287,7 +287,7 @@ impl LocalTestServer {
     ///
     /// The server listens on a randomly assigned available port on localhost.
     pub async fn start_with_socket_proxy(in_memory: bool, socket_path: Option<PathBuf>) -> Self {
-        let ctx = XetContext::default().expect("XetContext::new");
+        let ctx = XetRuntime::default().expect("XetRuntime::new");
         if in_memory {
             let client = MemoryClient::new(ctx.clone());
             Self::start_with_client_and_socket(ctx, client, None, socket_path).await
@@ -302,7 +302,7 @@ impl LocalTestServer {
     ///
     /// Useful when you need to pre-populate the client with data before starting the server.
     pub async fn start_with_client(client: Arc<dyn DirectAccessClient>) -> Self {
-        let ctx = XetContext::default().expect("XetContext::new");
+        let ctx = XetRuntime::default().expect("XetRuntime::new");
         Self::start_with_client_and_socket(ctx, client, None, None).await
     }
 
@@ -312,13 +312,13 @@ impl LocalTestServer {
         client: Arc<dyn DirectAccessClient>,
         deletion_client: Option<Arc<dyn DeletionControlableClient>>,
     ) -> Self {
-        let ctx = XetContext::default().expect("XetContext::new");
+        let ctx = XetRuntime::default().expect("XetRuntime::new");
         Self::start_with_client_and_socket(ctx, client, deletion_client, None).await
     }
 
     /// Starts a new test server using an existing `DirectAccessClient` with an optional socket proxy.
     async fn start_with_client_and_socket(
-        ctx: XetContext,
+        ctx: XetRuntime,
         client: Arc<dyn DirectAccessClient>,
         deletion_client: Option<Arc<dyn DeletionControlableClient>>,
         _socket_path: Option<PathBuf>,
@@ -1243,7 +1243,7 @@ mod tests {
     /// deletion controls.
     #[tokio::test]
     async fn test_deletion_lifecycle_via_server() {
-        let lc = LocalClient::temporary(XetContext::default().expect("ctx")).await.unwrap();
+        let lc = LocalClient::temporary(XetRuntime::default().expect("ctx")).await.unwrap();
         let server = LocalTestServer::start_with_client(lc.clone()).await;
 
         // Upload files via remote client (goes through HTTP server)
@@ -1318,7 +1318,7 @@ mod tests {
     #[cfg_attr(feature = "smoke-test", ignore)]
     async fn test_simulation_control_client_common_suite() {
         crate::cas_client::simulation::client_unit_testing::test_client_functionality(|| async {
-            let lc = LocalClient::temporary(XetContext::default().expect("ctx")).await.unwrap();
+            let lc = LocalClient::temporary(XetRuntime::default().expect("ctx")).await.unwrap();
             let dc: Arc<dyn DeletionControlableClient> = lc.clone();
             let server = LocalTestServer::start_with_client_and_deletion(lc, Some(dc)).await;
             let endpoint = detach_server(server);
@@ -1373,7 +1373,7 @@ mod tests {
     #[cfg_attr(feature = "smoke-test", ignore)]
     async fn test_simulation_control_client_deletion_suite() {
         crate::cas_client::simulation::deletion_unit_testing::test_deletion_functionality(|| async {
-            let lc = LocalClient::temporary(XetContext::default().expect("ctx")).await.unwrap();
+            let lc = LocalClient::temporary(XetRuntime::default().expect("ctx")).await.unwrap();
             let dc: Arc<dyn DeletionControlableClient> = lc.clone();
             let server = LocalTestServer::start_with_client_and_deletion(lc, Some(dc)).await;
             let endpoint = detach_server(server);
