+++ xet_core_structures/src/metadata_shard/session_directory.rs
@@ -7,7 +7,7 @@ use std::sync::Arc;
 use tokio::task::JoinHandle;
 use tracing::{error, info};
 use xet_runtime::RuntimeError;
-use xet_runtime::core::XetRuntime;
+use xet_runtime::core::XetThreadpool;
 
 use super::set_operations::shard_set_union;
 use super::shard_file_handle::MDBShardFile;
@@ -20,7 +20,7 @@ use crate::error::Result;
 ///
 /// Ordering of staged shards is preserved.
 pub fn consolidate_shards_in_directory(
-    runtime: &Arc<XetRuntime>,
+    runtime: &Arc<XetThreadpool>,
     session_directory: impl AsRef<Path>,
     target_max_size: u64,
     skip_on_error: bool,
@@ -57,7 +57,7 @@ pub struct ShardMergeResult {
 /// Ordering of staged shards is preserved.
 #[allow(clippy::needless_range_loop)] // The alternative is less readable IMO
 pub fn merge_shards(
-    runtime: &Arc<XetRuntime>,
+    runtime: &Arc<XetThreadpool>,
     source_directory: impl AsRef<Path>,
     target_directory: impl AsRef<Path>,
     target_max_size: u64,
@@ -158,7 +158,7 @@ pub fn merge_shards(
 
 /// Same as above, but performs it in the background and on a io focused thread.
 pub fn merge_shards_background(
-    runtime: Arc<XetRuntime>,
+    runtime: Arc<XetThreadpool>,
     source_directory: impl AsRef<Path>,
     target_directory: impl AsRef<Path>,
     target_max_size: u64,
