+++ xet_data/src/processing/file_download_session.rs
@@ -11,7 +11,7 @@ use tracing::instrument;
 use xet_client::cas_client::Client;
 use xet_client::cas_types::FileRange;
 use xet_client::chunk_cache::ChunkCache;
-use xet_runtime::core::XetContext;
+use xet_runtime::core::XetRuntime;
 
 use super::XetFileInfo;
 use super::configurations::TranslatorConfig;
@@ -25,7 +25,7 @@ use crate::progress_tracking::{GroupProgress, ItemProgressUpdater, UniqueID};
 /// This struct parallels `FileUploadSession` for the download path. It holds the
 /// CAS client and a shared progress group for all downloads in the session.
 pub struct FileDownloadSession {
-    ctx: XetContext,
+    ctx: XetRuntime,
     client: Arc<dyn Client>,
     chunk_cache: Option<Arc<dyn ChunkCache>>,
     progress: Arc<GroupProgress>,
@@ -65,7 +65,7 @@ impl FileDownloadSession {
     /// session should inherit the configured speed parameters from the context used
     /// to build [`TranslatorConfig`].
     pub fn from_client(
-        ctx: &XetContext,
+        ctx: &XetRuntime,
         client: Arc<dyn Client>,
         chunk_cache: Option<Arc<dyn ChunkCache>>,
     ) -> Arc<Self> {
@@ -119,7 +119,7 @@ impl FileDownloadSession {
         self.check_not_finalized()?;
         let id = UniqueID::new();
         let session = self.clone();
-        let rt = self.ctx.runtime.clone();
+        let rt = self.ctx.threadpool.clone();
         let semaphore = self.ctx.common.file_download_semaphore.clone();
         let handle = rt.spawn(async move {
             let _permit = semaphore.acquire().await?;
@@ -380,27 +380,27 @@ mod tests {
     use std::sync::{Arc, OnceLock};
 
     use tempfile::tempdir;
-    use xet_runtime::core::XetContext;
+    use xet_runtime::core::XetRuntime;
 
     use super::*;
     use crate::processing::configurations::TranslatorConfig;
     use crate::processing::file_cleaner::Sha256Policy;
     use crate::processing::{FileUploadSession, XetFileInfo};
 
-    fn get_threadpool() -> Arc<xet_runtime::core::XetRuntime> {
-        static THREADPOOL: OnceLock<Arc<xet_runtime::core::XetRuntime>> = OnceLock::new();
+    fn get_threadpool() -> Arc<xet_runtime::core::XetThreadpool> {
+        static THREADPOOL: OnceLock<Arc<xet_runtime::core::XetThreadpool>> = OnceLock::new();
         THREADPOOL
             .get_or_init(|| {
-                XetContext::default()
+                XetRuntime::default()
                     .expect("Error starting multithreaded runtime.")
-                    .runtime
+                    .threadpool
                     .clone()
             })
             .clone()
     }
 
     async fn upload_data(cas_path: &Path, data: &[u8]) -> XetFileInfo {
-        let ctx = XetContext::default().unwrap();
+        let ctx = XetRuntime::default().unwrap();
         let upload_session = FileUploadSession::new(TranslatorConfig::local_config(&ctx, cas_path).unwrap().into())
             .await
             .unwrap();
@@ -426,7 +426,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("output.txt");
@@ -450,7 +450,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("deep").join("nested").join("dir").join("output.txt");
@@ -475,7 +475,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("partial_writer.txt");
@@ -504,7 +504,7 @@ mod tests {
                 let original_data = b"abcdefghijklmnopqrstuvwxyz0123456789";
 
                 let xfi = upload_data(&cas_path, original_data).await;
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("partitioned.txt");
@@ -556,7 +556,7 @@ mod tests {
                 let xfi_a = upload_data(&cas_path, data_a).await;
                 let xfi_b = upload_data(&cas_path, data_b).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_a = temp.path().join("out_a.txt");
@@ -595,7 +595,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream(&xfi, None).await.unwrap();
@@ -622,7 +622,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, stream) = session.download_stream(&xfi, None).await.unwrap();
@@ -655,7 +655,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream(&xfi, None).await.unwrap();
@@ -684,7 +684,7 @@ mod tests {
                 let xfi_a = upload_data(&cas_path, data_a).await;
                 let xfi_b = upload_data(&cas_path, data_b).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id_a, mut stream_a) = session.download_stream(&xfi_a, None).await.unwrap();
@@ -727,7 +727,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, stream) = session.download_stream(&xfi, None).await.unwrap();
@@ -753,7 +753,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 for i in 0..5u32 {
@@ -784,7 +784,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, stream) = session.download_stream(&xfi, None).await.unwrap();
@@ -817,7 +817,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream(&xfi, None).await.unwrap();
@@ -840,7 +840,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream(&xfi, None).await.unwrap();
@@ -870,7 +870,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("range_from.bin");
@@ -895,7 +895,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("range_to.bin");
@@ -920,7 +920,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("full_range.bin");
@@ -945,7 +945,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("range_incl.bin");
@@ -972,7 +972,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream_range(&xfi, 4..12).await.unwrap();
@@ -999,7 +999,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream_range(&xfi, 10..).await.unwrap();
@@ -1026,7 +1026,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream_range(&xfi, ..6).await.unwrap();
@@ -1056,7 +1056,7 @@ mod tests {
                 let xfi = upload_data(&cas_path, original_data).await;
                 let xfi_no_size = XetFileInfo::new_hash_only(xfi.hash().to_string());
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("output_unknown.txt");
@@ -1081,7 +1081,7 @@ mod tests {
                 let xfi = upload_data(&cas_path, original_data).await;
                 let xfi_no_size = XetFileInfo::new_hash_only(xfi.hash().to_string());
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let (_id, mut stream) = session.download_stream(&xfi_no_size, None).await.unwrap();
@@ -1110,7 +1110,7 @@ mod tests {
                 let xfi = upload_data(&cas_path, original_data).await;
                 let wrong_size_xfi = XetFileInfo::new(xfi.hash().to_string(), 999);
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("output_mismatch.txt");
@@ -1157,7 +1157,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("empty_range.bin");
@@ -1182,7 +1182,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("inverted_range.bin");
@@ -1207,7 +1207,7 @@ mod tests {
 
                 let xfi = upload_data(&cas_path, original_data).await;
 
-                let config = TranslatorConfig::local_config(&XetContext::default().unwrap(), &cas_path).unwrap();
+                let config = TranslatorConfig::local_config(&XetRuntime::default().unwrap(), &cas_path).unwrap();
                 let session = FileDownloadSession::new(config.into(), None).await.unwrap();
 
                 let out_path = temp.path().join("beyond_size.bin");
