+++ xet_pkg/src/xet_session/session.rs
@@ -9,7 +9,7 @@ use ulid::Ulid;
 use xet_data::progress_tracking::UniqueID;
 use xet_runtime::RuntimeError;
 use xet_runtime::config::XetConfig;
-use xet_runtime::core::{XetContext, XetRuntime};
+use xet_runtime::core::XetRuntime;
 
 use super::download_stream_group::{DownloadStreamGroupBuilder, XetDownloadStreamGroup, XetDownloadStreamGroupInner};
 use super::errors::SessionError;
@@ -21,7 +21,7 @@ use super::upload_commit::{UploadCommitBuilder, XetUploadCommit};
 /// Lives behind `Arc<XetSessionInner>` — do not use this type directly.
 #[doc(hidden)]
 pub struct XetSessionInner {
-    pub(super) ctx: XetContext,
+    pub(super) ctx: XetRuntime,
 
     // CAS endpoint and shared HTTP settings (auth lives at the commit/group level)
     pub(super) endpoint: Option<String>,
@@ -184,20 +184,20 @@ impl XetSessionBuilder {
         let ctx = match handle {
             Some(h) => {
                 info!("XetSession using External runtime (wrapping caller's tokio handle)");
-                match XetRuntime::from_external_with_config(h, &self.config) {
-                    Ok(runtime) => XetContext::new(runtime, self.config.clone()),
+                match XetRuntime::from_external_with_config(h, self.config.clone()) {
+                    Ok(ctx) => ctx,
                     Err(RuntimeError::ExternalAlreadyAttached(_)) => {
                         info!(
                             "An existing XetSession already wraps caller's tokio handle, switching to creating Owned runtime"
                         );
-                        XetContext::default_with_config(self.config.clone())?
+                        XetRuntime::default_with_config(self.config.clone())?
                     },
                     Err(e) => Err(e)?,
                 }
             },
             None => {
                 info!("XetSession creating Owned runtime (new thread pool)");
-                XetContext::default_with_config(self.config.clone())?
+                XetRuntime::default_with_config(self.config.clone())?
             },
         };
 
@@ -238,8 +238,8 @@ pub struct XetSession {
 
 impl XetSession {
     /// Low-level constructor used by [`XetSessionBuilder::build`].
-    fn new(endpoint: Option<String>, custom_headers: Option<Arc<HeaderMap>>, ctx: XetContext) -> Self {
-        let task_runtime = TaskRuntime::new_root(ctx.runtime.clone());
+    fn new(endpoint: Option<String>, custom_headers: Option<Arc<HeaderMap>>, ctx: XetRuntime) -> Self {
+        let task_runtime = TaskRuntime::new_root(ctx.threadpool.clone());
         Self {
             inner: Arc::new(XetSessionInner {
                 ctx,
@@ -332,7 +332,7 @@ impl XetSession {
     /// This does not call per-commit/group local abort hooks.
     pub fn sigint_abort(&self) -> Result<(), SessionError> {
         info!("Session SIGINT abort, session_id={}", self.inner.id);
-        self.inner.ctx.runtime.perform_sigint_shutdown();
+        self.inner.ctx.threadpool.perform_sigint_shutdown();
 
         self.inner.active_upload_commits.lock()?.clear();
         self.inner.active_file_download_groups.lock()?.clear();
@@ -348,7 +348,7 @@ impl XetSession {
 
     #[cfg(test)]
     pub(super) fn check_alive(&self) -> Result<(), SessionError> {
-        if self.inner.ctx.runtime.in_sigint_shutdown() {
+        if self.inner.ctx.threadpool.in_sigint_shutdown() {
             return Err(SessionError::KeyboardInterrupt);
         }
         self.inner.task_runtime.check_state("session")
@@ -649,7 +649,7 @@ mod tests {
     // build_blocking on an External-mode session returns WrongRuntimeMode.
     async fn test_new_upload_commit_blocking_errors_in_external_mode() {
         let session = XetSessionBuilder::new().build().unwrap();
-        assert_eq!(session.inner.ctx.runtime.mode(), RuntimeMode::External);
+        assert_eq!(session.inner.ctx.threadpool.mode(), RuntimeMode::External);
         let err = session.new_upload_commit().unwrap().build_blocking().err().unwrap();
         assert!(matches!(err, SessionError::WrongRuntimeMode(_)));
     }
@@ -658,7 +658,7 @@ mod tests {
     // build_blocking on an External-mode session returns WrongRuntimeMode.
     async fn test_new_file_download_group_blocking_errors_in_external_mode() {
         let session = XetSessionBuilder::new().build().unwrap();
-        assert_eq!(session.inner.ctx.runtime.mode(), RuntimeMode::External);
+        assert_eq!(session.inner.ctx.threadpool.mode(), RuntimeMode::External);
         let err = session.new_file_download_group().unwrap().build_blocking().err().unwrap();
         assert!(matches!(err, SessionError::WrongRuntimeMode(_)));
     }
@@ -669,7 +669,7 @@ mod tests {
     // build_blocking panics when called from within a tokio runtime on an Owned-mode session.
     fn test_new_upload_commit_blocking_panics_in_async_context() {
         let session = XetSessionBuilder::new().build().unwrap();
-        assert_eq!(session.inner.ctx.runtime.mode(), RuntimeMode::Owned);
+        assert_eq!(session.inner.ctx.threadpool.mode(), RuntimeMode::Owned);
         let rt = tokio::runtime::Builder::new_multi_thread().enable_all().build().unwrap();
         let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
             rt.block_on(async { session.new_upload_commit().unwrap().build_blocking() })
@@ -681,7 +681,7 @@ mod tests {
     // build_blocking panics when called from within a tokio runtime on an Owned-mode session.
     fn test_new_file_download_group_blocking_panics_in_async_context() {
         let session = XetSessionBuilder::new().build().unwrap();
-        assert_eq!(session.inner.ctx.runtime.mode(), RuntimeMode::Owned);
+        assert_eq!(session.inner.ctx.threadpool.mode(), RuntimeMode::Owned);
         let rt = tokio::runtime::Builder::new_multi_thread().enable_all().build().unwrap();
         let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
             rt.block_on(async { session.new_file_download_group().unwrap().build_blocking() })
@@ -886,12 +886,12 @@ mod tests {
         let handle = tokio_rt.handle().clone();
 
         let first = XetSessionBuilder::new().with_tokio_handle(handle.clone()).build().unwrap();
-        assert_eq!(first.inner.ctx.runtime.mode(), RuntimeMode::External, "first build must use External runtime");
+        assert_eq!(first.inner.ctx.threadpool.mode(), RuntimeMode::External, "first build must use External runtime");
 
         let second = XetSessionBuilder::new().with_tokio_handle(handle).build();
         assert!(second.is_ok(), "second build with the same tokio handle must still succeed");
         assert_eq!(
-            second.unwrap().inner.ctx.runtime.mode(),
+            second.unwrap().inner.ctx.threadpool.mode(),
             RuntimeMode::Owned,
             "second build must fall back to Owned runtime when External handle is already in use"
         );
