+++ xet_runtime/src/error.rs
@@ -9,7 +9,7 @@ pub enum RuntimeError {
     #[error("Invalid runtime: {0}")]
     InvalidRuntime(String),
 
-    #[error("A XetRuntime is already attached to this tokio runtime handle with Id {0}")]
+    #[error("A XetThreadpool is already attached to this tokio runtime handle with Id {0}")]
     ExternalAlreadyAttached(tokio::runtime::Id),
 
     #[error("Task panic: {0:?}")]
